requirements.txt
python-dotenv==1.1.0
google-api-python-client==2.124.0
google-auth==2.29.0
google_auth_oauthlib==1.2.0
